# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: SHIELD API
  version: 1.0.0
  description: Smart Housing Infrastructure and Entry Log Digitalization API
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /complaints:
    post:
      tags: [Complaint]
      summary: Create complaint
      responses:
        '200': {description: Complaint created}
    get:
      tags: [Complaint]
      summary: List complaints
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': {description: Complaints fetched}
  /complaints/{id}:
    get:
      tags: [Complaint]
      summary: Get complaint by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Complaint fetched}
    put:
      tags: [Complaint]
      summary: Update complaint
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Complaint updated}
    delete:
      tags: [Complaint]
      summary: Delete complaint
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Complaint deleted}
  /complaints/{id}/assign:
    put:
      tags: [Complaint]
      summary: Assign complaint
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Complaint assigned}
    post:
      tags: [Complaint]
      summary: Assign complaint (alias)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Complaint assigned}
  /complaints/{id}/resolve:
    put:
      tags: [Complaint]
      summary: Resolve complaint
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Complaint resolved}
    post:
      tags: [Complaint]
      summary: Resolve complaint (alias)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Complaint resolved}
  /complaints/{id}/close:
    post:
      tags: [Complaint]
      summary: Close complaint
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Complaint closed}
  /complaints/{id}/reopen:
    post:
      tags: [Complaint]
      summary: Reopen complaint
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Complaint reopened}
  /complaints/status/{status}:
    get:
      tags: [Complaint]
      summary: List complaints by status
      parameters:
        - in: path
          name: status
          required: true
          schema: {type: string}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Complaints by status fetched}
  /complaints/priority/{priority}:
    get:
      tags: [Complaint]
      summary: List complaints by priority
      parameters:
        - in: path
          name: priority
          required: true
          schema: {type: string}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Complaints by priority fetched}
  /complaints/asset/{assetId}:
    get:
      tags: [Complaint]
      summary: List complaints by asset
      parameters:
        - in: path
          name: assetId
          required: true
          schema: {type: string, format: uuid}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Complaints by asset fetched}
  /complaints/my-complaints:
    get:
      tags: [Complaint]
      summary: List my complaints
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: My complaints fetched}
  /complaints/assigned-to-me:
    get:
      tags: [Complaint]
      summary: List complaints assigned to current user
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Assigned complaints fetched}
  /complaints/sla-breached:
    get:
      tags: [Complaint]
      summary: List SLA breached complaints
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: SLA breached complaints fetched}
...
  /complaints/{id}/comments:
    get:
      tags: [Complaint]
      summary: List complaint comments
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Complaint comments fetched}
    post:
      tags: [Complaint]
      summary: Add complaint comment
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Complaint comment added}
...
  /work-orders/complaint/{complaintId}:
    get:
      tags: [Complaint]
      summary: List work orders by complaint
      parameters:
        - in: path
          name: complaintId
          required: true
          schema: {type: string, format: uuid}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Work orders by complaint fetched}
...
  /fire-drill-records:
    get:
      tags: [Emergency]
      summary: List fire drill records
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': {description: Fire drill records fetched}
    post:
      tags: [Emergency]
      summary: Create fire drill record
      responses:
        '200': {description: Fire drill record created}
  /fire-drill-records/{id}:
    get:
      tags: [Emergency]
      summary: Get fire drill record
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Fire drill record fetched}
    put:
      tags: [Emergency]
      summary: Update fire drill record
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Fire drill record updated}
    delete:
      tags: [Emergency]
      summary: Delete fire drill record
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Fire drill record deleted}
  /fire-drill-records/year/{year}:
    get:
      tags: [Emergency]
      summary: List fire drill records by year
      parameters:
        - in: path
          name: year
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': {description: Fire drill records by year fetched}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      in: query
      name: size
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    Sort:
      in: query
      name: sort
      schema:
        type: string
        example: createdAt,desc
...